# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Pick_ErrorResponsePayload.Exclude_keyofErrorResponsePayload.details-or-stack__": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"success",
					"message"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ErrorResponsePayload.details-or-stack_": {
				"$ref": "#/components/schemas/Pick_ErrorResponsePayload.Exclude_keyofErrorResponsePayload.details-or-stack__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CommonErrorPayload": {
				"$ref": "#/components/schemas/Omit_ErrorResponsePayload.details-or-stack_"
			},
...
			"GetCategoriesResDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"secondCategories": {
						"items": {
							"properties": {
								"name": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"name",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"secondCategories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponsePayload_GetCategoriesResDto-Array_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetCategoriesResDto"
						},
						"type": "array"
					}
				},
				"required": [
					"statusCode",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponsePayload_null_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryReqDto": {
				"properties": {
					"firstCategory": {
						"type": "string"
					},
					"secondCategories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"firstCategory",
					"secondCategories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"Pick_ErrorResponsePayload.Exclude_keyofErrorResponsePayload.stack__": {
				"properties": {
					"details": {
						"$ref": "#/components/schemas/FieldErrors"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"success",
					"message"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ErrorResponsePayload.stack_": {
				"$ref": "#/components/schemas/Pick_ErrorResponsePayload.Exclude_keyofErrorResponsePayload.stack__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ValidationErrorPayload": {
				"$ref": "#/components/schemas/Omit_ErrorResponsePayload.stack_"
			},
			"CreateSecondCategoryReqDto": {
				"properties": {
					"secondCategory": {
						"type": "string"
					}
				},
				"required": [
					"secondCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponsePayload": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"details": {
						"$ref": "#/components/schemas/FieldErrors"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"description": "toBeDev Project Server",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"email": "pca0046@gmail.com",
			"name": "박채인"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePayload_GetCategoriesResDto-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "PostCategory",
				"responses": {
					"201": {
						"description": "CREATED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePayload_null_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorPayload"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonErrorPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryReqDto"
							}
						}
					}
				}
			}
		},
		"/categories/{firstCategoryId}/second-categories": {
			"post": {
				"operationId": "PostSecondCategory",
				"responses": {
					"201": {
						"description": "CREATED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePayload_null_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorPayload"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponsePayload"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "firstCategoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSecondCategoryReqDto"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8080/"
		}
	]
}